requirements.txt
google_cloud_speech
google_cloud_storage
pandas==1.3.0
